Check if the renaming folder already exists

+++ Engine/Editor/AssetsExplorer/AssetsExplorer.cpp
@@ -217,11 +217,15 @@ void AssetsExplorer() {
                 if (fs::exists(rename_folder_name)) {
                     fs::path newFolderPath = rename_folder_name.parent_path() / rename_folder_buffer;
                     try {
-                        fs::rename(rename_folder_name, newFolderPath);
+                        if (fs::exists(newFolderPath)) {
+                            std::cerr << "Target directory already exists: " << newFolderPath.string() << std::endl;
+                        } else {
+                            fs::rename(rename_folder_name, newFolderPath);
+                            std::cout << "Folder renamed successfully to " << newFolderPath.string() << std::endl;
+                        }
                     } catch (const std::filesystem::filesystem_error& e) {
                         std::cerr << "Error renaming folder: " << e.what() << std::endl;
                     }
-                    std::cout << "Folder renamed successfully to " << newFolderPath.string() << std::endl;
                 } else {
                     std::cout << "Folder not found: " << rename_folder_name.string() << std::endl;
                 }
